cmd/cue: remove fallback version code for Go 1.17
As of https://cuelang.org/cl/545975 in November 2022,
we now require at least Go 1.18 when building CUE.
