+++ cmd/cue/cmd/testdata/script/version.txtar
@@ -8,10 +8,9 @@ stdout -count=1 'cue version \(devel\)'
 # Version output with VCS information.
 env CUE_VERSION_TEST_CFG='[{"Key":"vcs","Value":"git"},{"Key":"vcs.revision","Value":"47b7032385cb490fab7d47b89fca36835cf13d39"},{"Key":"vcs.time","Value":"2022-05-10T04:58:46Z"},{"Key":"vcs.modified","Value":"true"}]'
 exec cue version
-[!go1.18] stdout -count=1 '^cue version \(devel\) [a-z]+/[a-z]+'
-[go1.18] stdout -count=1 '^cue version v0.0.0-20220510045846-47b7032385cb$'
-[go1.18] stdout -count=1 '^go version (devel )?go1.'
-[go1.18] stdout -count=1 'vcs git'
-[go1.18] stdout -count=1 'vcs\.revision 47b7032385cb490fab7d47b89fca36835cf13d39'
-[go1.18] stdout -count=1 'vcs\.time 2022-05-10T04:58:46Z'
-[go1.18] stdout -count=1 'vcs\.modified true'
+stdout -count=1 '^cue version v0.0.0-20220510045846-47b7032385cb$'
+stdout -count=1 '^go version (devel )?go1.'
+stdout -count=1 'vcs git'
+stdout -count=1 'vcs\.revision 47b7032385cb490fab7d47b89fca36835cf13d39'
+stdout -count=1 'vcs\.time 2022-05-10T04:58:46Z'
+stdout -count=1 'vcs\.modified true'
